fix: correctly parse git status
I initially assumed that a line resulting of the `git status --porcelain=2 --branch` command like the following,

```
 M test.txt
```

means that it is modified and unstaged and

```
M  test.txt
```

means that it is modified and staged.

I then assumed that I would just get two lines when a change to a file
is unstaged and a change to the same file is staged, but it looks like
you just get a line like the following.

```
MM test.txt
```