fix: process cb now waits for all pipes to close
The on-exit callback being called by libuv's spawn does not indicate
that all data has already been sent. We now wait not only for the
process to close, but also for all pipes to signal they have sent all
the data.
Only after all output pipes and the process are closed, the response
callback is invoked with the available data.