refactor: better way to build cli calls
Add a git command builder to easily assemble git commands and call them.
The available flags and options are configured via a central
`configuration` object and can then be added via a simple fluent
pattern, e.g.:

cli.status
  .branch
  .porcelain(1)
  .call()

The call method uses the async module and needs to be awaited.
Additionally, there's a utility function `in_parallel` which takes any
number of command builders and executes them in parallel, making sure
the cwd is set for each of them.