feat: better fold-keeping
Refactor of the status buffer rendering algorithm to help vim keep track
of folds. Instead of re-writing the complete buffer from scratch, the
buffer is now incrementally rendered, making sure only necessary data
changes in the buffer.

This helps vim keep track of existing folds. As long as vim can track
the fold, removing or adding content to it will generally keep the fold
open, as long as it was opened before.

I observed a few strange edge-cases tough:
- Vim seems to set a fold's initial state based on it's surrounding
  fold's state; if a fold is already open, a new fold created below it
  at the same level will usually also be open.
- On the other hand though, vim will restore the folds to the actual
  foldlevel again on occasion, for example when moving down the last
  part in a section