fix(push): fix push elsewhere
The original function contains two bugs. First is that it didn't pass
popup arguments. Because of this, no state was found when refreshing the
status.

Secondly, the git.branch.prompt_for_branch API needs a list of
branches to test if arguments is contained in options. But the original
code just call it without any arguments. I don't know what arguments
should be passed into it.  Keep using input.get_user_input is my current
workaround.

Signed-off-by: Avimitin <avimitin@gmail.com>