refactor(push): allow user to input custom branch.. ...name which is not exist in git status
When using the elsewhere action, user have to use the selection in the
auto-completion. Or they will get nil error if their input is not in the
given selections.

This is not the behavior I expect. As what we talk about in the
discussion in PR #233, the push elsewhere action will try to fetch all
the local branches. Those branches all have long and not useful name.
For user who want to push to remote with name gitea and branch with name
some-branch, the prompt will force them to select
"remotes/gitea/some-branch". For users, maybe they only want to input
"some-branch". Besides, the command "git push gitea
remotes/gitea/some-branch" is also an illegal CLI argument.

Although I've already truncate the branch argument (See commit 12f8e1669),
I don't think this is the best workaround. We should allow user input
what they want just like they are using the git CLI. The auto completion
should be a helper not a limit.

Signed-off-by: Avimitin <avimitin@gmail.com>