feat(push): get tracking upstream before push
The behavior for the push to upstream action is not what I expected.
When I press u it actually runs the command "git push upstream branch".
But what I expect is it can push to user-defined upstream. Which is set
by the "git push -u" command.

So here I use the command "git branch -vv" to get the tracking remote
information. The output of the tracking upstream will have a "*" prefix.
The upstream information is wrapped in the "[]" signs. For example, in
an output like this: "* master 3d9279e [github/master] docs: update title".
The "*" mark this line as the tracking upstream. And the upstream remote
is GitHub, the branch is master.

I add a new flag "-vv" in lua/neogit/lib/git/cli.configuration.branch.
And call the cli.branch.tracking.call() function to get the branch text.
Besides I implement the get_upstream() function in lua/neogit/lib/git/branch,
it matches the string and returns a table with the remote name and
tracking upstream name. It will return nil when no match is found. If
the action detected nil return, it will call the logger to print an
error and return.

Signed-off-by: Avimitin <avimitin@gmail.com>