Further optimize diff generation. I was encountering an issue where, with a large number of files in the index neogit would crash, with an error complaining of too many open file descriptors.
To rememdy this, I refactored the metamethod logic to not just be for
the hunks, but for the _entire_ diff. Importantly, this includes the cli
calls to get the raw diff/stats. Further, calls to check the _presence_
of a diff would trigger these functions, so we add a 'has_diff' key to
provide an inexpensive way to check for the ability to do the more
expensive call later on.

Lastly, I think it's fine to do a blocking call, since we need the
result of the function before rerendering the status buffer. in my
testing it's not noticable with a 100,000 line file.