refactor!: improve mapping setup
This makes it easier to define which command goes to a key. Prior to
this you'd have to set a key to empty like so: `["s"] = ""` then set
another key to change the binding.

Now the binding is set to `Command -> Key` instead of `Key -> Command`.
This also fixes some bugs in which a user sets two bindings to empty to
remove them which currently causes neogit to crash.