refactor! remove magit mappings
This is hardly utilized, and it is suggested to override the mappings
yourself to the way you want it. Additionally, the meaning is lost as the DOOM mappings are
different from Vanilla Magit mapping, and vanilla mappings use mappings
which are not compatible with Vim, such as `C-c C` etc.