fix: follow commit under cursor in log buffer
This didn't work as it tried to vim.search for the full oid, which isn't
necessarily what git and our ui will display. This uses the more robust
approach of acessing the ui stack rather than searching, which won't be
mislead by ui formatting such as slashes, abbreviations, and commit ids
in commit messages etc.