Bugfix: Rewrite how parsing commits works.
When a commit message contained a newline, it would print on another
line in the output. We parsed this as a "graph" line, and it would
result in strange lines on the commit log buffer.

The fix is to rewrite how the commits are parsed. First, we insert a
Terminator character at the end of each commit entry, to split on later.

Then we concat all the lines from the output with newline characters, to
preserve formatting. Next, we split on the character we inserted
earlier, '\31'. Finally, we map over our commits, replacing newline
characters with spaces, and splitting on our delimiter, '\30'.

This leaves us with a table, where each item is a single commit entry
sans any graph information.

The next change is that we need to now re-sync the entries with the
graph. To do that, we are passing in the un-parsed graph as a table
split on newlines. We can check each line, and if there is a '*', then a
commit entry goes on that line. Otherwise, it's just a graph line
without commit.