feat: Show tag name and distance from head on status page
Similar to Magit's status page this commit adds the most recent tag name
to the status popup with the distance (commits) between the head and tag

Unfortunately lua's pattern matching isn't as complete as I had hoped
for so there might be some cases where it fails. When it fails it will
just set the tag and distance to nil.