Parse `tag: ...` and render with `NeogitTagName` in Log Graph
The `commit.ref_name` might also contain a tag name, which the log graph
is currently rendering in the `NeogitBranch` highlight group. The
refname of HEAD with a tag placed on it might look like this:

> HEAD -> main, tag: 1.0.0, origin/main, origin/HEAD

A refname of a tag not associated to any branch might look like this:

> tag: 1.0.0

This makes the parsing of `local_name`, `remote_name` a bit more
sophisticated, since we don't know the exact order of them. Naive
solution might be to just split `ref_name` by comma and pattern match:

* contains `tag: ...` -> `tag_name`
* contains `/` -> `remote_name`
* otherwise -> `local_name`