Properly interprete branch names, remotes & tags
Parsing the the `ref_name` manually by regex is finicky and error prone.
To better identify, which of the parts of `ref_name` are remotes and
which are branch names one can not rely on finding the `/`, since
branches might also contain `/`. Also the order of the fields within the
`ref_name` can vary.

For a robust solution I extracted this parsing into the
`lib/git/log.lua` module and added unit tests as well. This simplifies
the logic of the logging function as well.