Fix ordering of tracked/untracked branches in log graph
Since lua tables are not sorted, the exact order in which the branch
names are placed in the log graph can vary and depdends on the output of
`git log...`. For example in Git version 2.34 it would:

> e2c2a1c * origin/tracked-branch untracked-branch MESSAGE

whereas in Git 2.39 it would be reversed:

> e2c2a1c * untracked-branch origin/tracked-branch MESSAGE

To keep the behaviour consistent between git versions, I introduced the
notion of `tracked` and `untracked` branches in the
`git.log.branch_info()`.

* `untracked`: List of all branches which have no remote counterpart
  (i.e. are local only)
* `tracked`: Mapping of local branch names (keys) which a list of remote
  names (values) which contain such branches.