Use ; and , to scroll commit view inside log view
When viewing the commit log, pressing `;` or `,` can be used to open the
commit view without focusing it. If you press `;` again, it will scroll
down the commit view. With `,` you can scroll up. Because the commit
view window isn't focused, you can quickly focus the next commit with
`j` or the previous with `k`.

This reduces the amount of key presses significantly when reviewing
changes of multiple commits because you don't have to switch windows
twice for every commit you want to look at.

Since `f` and `t` are bound differently to normal vim usage, `;` and `,`
are also basically free to use. On the US layout (and others), you only
need the right hand to review the changes and can use the left hand to
drink coffee or tee. ;-)

These bindings implement the same functionality as
`magit-diff-show-or-scroll-down` and `magit-diff-show-or-scroll-up`.