Add config feature 'initial_branch_name'
This config feature allows users to specify an initial branch name that
will be used as a `get_user_input` default when Neogit asks for a name
for a new branch.

This is useful when working in teams with a branch naming convention.
For example, teams may choose to prefix branches with a GitHub
username. In that case, configuring Neogit with:

```
initial_branch_name = "myname/"
```

...will result in the popup being prepopulated with `myname/`, after
which I can specify the branch name.