Fix error when trying to diff worktree
Error is:
.../neogit/integrations/diffview.lua:165: attempt to concatenate local 'item_name' (a nil value)

It looks like `diffview.open` is called by the worktree function in
`popups/diff/actions.lua`:104 with no arguments, causing item_name to
be nil.

This fix checks to see if item_name is null and, if so, just calls
`dv_lib.diffview_open` with no arguments as indicated by the DiffView
docs.