fix: process.lua error when we have a bare repo in the worktree list
When generating a list of worktrees using the 'git worktree list --porcelain'
command, a bare worktree only has 2 lines, whereas all other worktrees have 3
lines. Since we don't need to switch to the bare repo, we can safely ignore
it. The new listing function will look for a branch listing and only add that
to the list of worktrees.