feat: Add auto_show_console_on = "output" | "error" setting
This pairs very well with the new spinner feature:
https://github.com/NeogitOrg/neogit/pull/1551

Allows users to set:

```lua
{ auto_show_console_on = "error" }
```

Note that this is only considered if `auto_show_console` is `true` (this
is the default).

This results in Neogit showing a spinner while, e.g. pre-commit hooks
are running, and only shows the console if the pre-commit hooks fail.

The default is `auto_show_console_on = "output"`, which matches the
previous behavior.

This PR also stops Neogit from calling `notification.warn` on a bad exit
code, because this overlaps with showing the console, and requires the
user to press enter unnecessarily in order to get to the console output.
The `notification.warn` call is preserved if `auto_show_console` is
`false`.