Make spinner step messages transient
Using `echom` pushes the messages into the message history, which means
that if the git operation is long enough, we'll pollute the history with
redundant messages.

For some context, those messages are useful to explore the output of
different processes.

You can explore the history using the built-in `:messages` command, or
you can use something like `:Message` from vim-scriptease which sends
the message history into the quickfix window.

Since it would be great to keep at least one entry in the history, in
this commit I'm adding one entry to the messages history, prefixed with
`[neogit]` and then replace the spinner step messages with `echo`, which
also prints the message but it doesn't save it in the history.